Location update is now deactivated when activity is in background
(AND06D Aufg.4.2.) In order to save battery, the location service is
now deactivated, each time that the methods onStop and onPause are
called by android. The service is started again when the activity is
visible again, but only in the case that the location was active before
going to the background.